+++ src/script/bindings/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod helpers;
 pub mod http;
+pub mod ping;
 pub mod tcp;